+++ .github/workflows/deploy-preview.yml
@@ -1,0 +1,30 @@
+name: CI
+
+on:
+  push:
+    branches-ignore:
+      # Renovate branches are always Pull Requests.
+      # We don't need to run CI twice (push+pull_request)
+      - 'renovate/**'
+  pull_request:
+    paths:
+      - 'docs/**'
+
+permissions: {}
+
+jobs:
+  test-dev:
+    if: ${{ github.actor != 'l10nbot' }}
+    runs-on: ubuntu-latest
+    steps:
+      - name: Use Node.js 14.x
+        uses: actions/setup-node@8c91899e586c5b171469028077307d293428b516 # v3.5.1
+        with:
+          node-version: 14
+          cache: 'yarn' # https://github.com/actions/setup-node/blob/main/docs/advanced-usage.md#caching-packages-dependencies
+      - run: yarn install
+        env:
+          # Don't need playwright in this job
+          PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD: 1
+      - name: Add deploy previews to PR's comment
+        run: yarn danger ci -d dangerfile.deploy-preview.ts -i deploy-preview
